[File&Code]: 파일이름 변경, 추가 & 코드 수정
- 할일 상세목록에 제목, 설명과 타이머가 있는데
  이를 파일로 구분하고 컴포넌트화 하기위해 파일 추가 및 수정함
* TaskTimer.swift
* TaskProcessView.swift
- TaskDetail 뷰 코드 간단 수정
